backend/requirements.txt
joblib==1.4.2
numpy==2.2.0
pandas==2.2.3
pytest==8.3.5
scikit-learn==1.6.1
typing_extensions==4.15.0
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.44
